\name{get_malaria_incidence_peru}
\alias{get_malaria_incidence_peru}
\title{Get Tutorial Malaria Incidence Data}
\usage{
get_malaria_incidence_peru()
}
\description{
Load the observed monthly malaria incidence in Padre Cocha, Loreto, Peru, from the beginning of the year 1996 to the end of the year 1998.

Derived from Pizzitutti, F. et al. A validated agent-based model to study the spatial and temporal heterogeneities of malaria incidence in the rainforest environment. Malar. J. 14, 514 (2015).
[Figure 6]
}
